Versione 1.5 -Variabile child globale; -Creazione keyManager; -Rimuove dati più vecchi di 5 minuti; -Lo fa ogni 30 secondi; -INSERIRE SEMAFORI (DA FARE...)

+++ clientReq-server/src/server.c
@@ -1,3 +1,5 @@
+#define MEMORY_SIZE 10
+
 #include <stdlib.h>
 #include <stdio.h>
 
@@ -18,10 +20,12 @@ char *pathToServerFIFO = "/tmp/serverFIFO";
 char *baseClientFIFO = "/tmp/clientFIFO.";
 
 int serverFIFO, serverFIFO_extra;
+pid_t child;
 int stampa = 1;
 int salva = -2;
 int invia = 0;
 
+//FUNZIONI
 void signalManager(void){
   sigset_t set;
   if(sigfillset(&set) == -1)
@@ -33,14 +37,18 @@ void signalManager(void){
 }
 
 void quit(int sig){
-  if (sig == SIGTERM)
+  if (sig == SIGTERM){
+    kill(child, SIGTERM);
     printf("<Server> Terminazione...\n");
+  }
   if(serverFIFO != 0 && close(serverFIFO) == -1)
     errExit("close failed");
   if(serverFIFO_extra != 0 && close(serverFIFO_extra) == -1)
     errExit("close failed");
-  if (unlink(pathToServerFIFO) != 0)
-    errExit("unlink failed");
+  if (child != 0) {
+    if (unlink(pathToServerFIFO) != 0)
+      errExit("unlink failed");
+  }
   _exit(0);
 }
 
@@ -78,62 +86,91 @@ void sendResponse(struct Request *request, struct Response *response){
     printf("close failed\n");
 }
 
+//MAIN
+
 int main (int argc, char *argv[]) {
-    printf("Hi, I'm Server program!\n");
-    //inizio scrittura
-    signalManager(); //blocca tutti segnali tranne SIGTERM
-    printf("<Server> Tutti segnali bloccati tranne SIGTERM (int 15)\n\n");
-    //Memoria condivisa che salva fino a 10 richieste
-    int contatore = 0;
-    int n = 10;
-    size_t sizeMemory = sizeof(struct keyManager) * n;
-    int shmid = alloc_shared_memory(IPC_PRIVATE, sizeMemory);
-    //attach shared memory
-    struct keyManager *km = (struct keyManager *)get_shared_memory(shmid, S_IRUSR | S_IWUSR);
-    //creazione serverFIFO
-    if (mkfifo(pathToServerFIFO, S_IRUSR | S_IWUSR | S_IWGRP) == -1)
-      errExit("mkfifo failed");
-    printf("<Server> serverFIFO creata con successo...\n");
-    if (signal(SIGTERM, quit) == SIG_ERR)
-      errExit("signal failed");
-    serverFIFO = open(pathToServerFIFO, O_RDONLY);
-    //apertura FIFO server in sola scrittura così che non veda la fine del file...
-    if (serverFIFO == -1)
-      errExit("open failed");
-    serverFIFO_extra = open(pathToServerFIFO, O_WRONLY);
-    if (serverFIFO_extra == -1)
-      errExit("open failed");
-    struct Request request;
-    ssize_t bR = -1;
-    do {
-      printf("<Server> Attesa richieste...\n");
-      bR = read(serverFIFO, &request, sizeof(request));
-      if (bR == -1)
-        printf("<Server>FIFO non funzionante...\n");
-      else if(bR != sizeof(struct Request) || bR == 0)
-        printf("<Server> Non ho ricevuto una richiesta valida...\n");
-      else{
-        printf("<Server> Lettura Request eseguita con successo...\n");
-        struct Response response;
-        keyCreation(&request, &response);
-        if(response.key == -1)
-          printf("<Server> Servizio non disponibile...\n");
-        else{
-          strcpy((km+contatore)->identificativo, request.identificativo);
-          (km+contatore)->key = response.key;
-          (km+contatore)->timestamp = time(NULL);
-          contatore++;
-          if (contatore >= n)
-            contatore = 0;
-          //stampa shared_memory
-          printf("SHARED MEMORY\n");
-          for (int i = 0; i < n; i++) {
-            printf("%i) %s\n", i, (km+i)->identificativo);
-            printf("%i) %i\n", i, (km+i)->key);
-            printf("%i) %lu\n", i, (km+i)->timestamp);
+  printf("Hi, I'm Server program!\n");
+  //inizio scrittura
+  signalManager(); //blocca tutti segnali tranne SIGTERM
+  printf("<Server> Tutti segnali bloccati tranne SIGTERM (int 15)\n\n");
+  //Memoria condivisa
+  int contatore = 0;
+  size_t sizeMemory = sizeof(struct keyManager) * MEMORY_SIZE;
+  int shmid = alloc_shared_memory(IPC_PRIVATE, sizeMemory);
+  //attach shared memory
+  struct keyManager *km = (struct keyManager *)get_shared_memory(shmid, S_IRUSR | S_IWUSR);
+
+  //PROCESSO FIGLIO CHE CONTROLLA SHARED MEMORY OGNI 30 SECONDI
+  child = fork();
+  if (child == -1) {
+    printf("<Server> keyManager rotto\n");
+    errExit("fork failed");
+  }
+  if (child == 0) {
+    while (1) {
+      printf("<KeyManager> Dormo...\n");
+      sleep(30);
+      printf("<KeyManager> Sveglio...\n");
+      for (int i = 0; i < MEMORY_SIZE; i++) {
+        if (strlen((km+i)->identificativo) != 0 && (km+i)->key != 0 && (km+i)->timestamp != 0) {
+          if (difftime(time(NULL), (km+i)->timestamp) > 300) {
+            strcpy((km+i)->identificativo, "\0");
+            (km+i)->key = 0;
+            (km+i)->timestamp = 0;
+            printf("<KeyManager> Elimanta entry %i...\n", i+1);
           }
         }
-        sendResponse(&request, &response);
       }
-    } while(bR != -1);
+    }
+  }
+  //creazione serverFIFO
+  if (mkfifo(pathToServerFIFO, S_IRUSR | S_IWUSR | S_IWGRP) == -1)
+    errExit("mkfifo failed");
+  printf("<Server> serverFIFO creata con successo...\n");
+  if (signal(SIGTERM, quit) == SIG_ERR)
+    errExit("signal failed");
+  serverFIFO = open(pathToServerFIFO, O_RDONLY);
+  //apertura FIFO server in sola scrittura così che non veda la fine del file...
+  if (serverFIFO == -1)
+    errExit("open failed");
+  serverFIFO_extra = open(pathToServerFIFO, O_WRONLY);
+  if (serverFIFO_extra == -1)
+    errExit("open failed");
+  struct Request request;
+  ssize_t bR = -1;
+  do {
+    printf("<Server> Attesa richieste...\n");
+    bR = read(serverFIFO, &request, sizeof(request));
+    if (bR == -1)
+    printf("<Server>FIFO non funzionante...\n");
+    else if(bR != sizeof(struct Request) || bR == 0)
+    printf("<Server> Non ho ricevuto una richiesta valida...\n");
+    else{
+      printf("<Server> Lettura Request eseguita con successo...\n");
+      struct Response response;
+      keyCreation(&request, &response);
+      if(response.key == -1)
+      printf("<Server> Servizio non disponibile...\n");
+      else{
+        //INSERIMENTO VALORI IN SHARED MEMORY
+        printf("<Server> Contatore: %i\n", contatore);
+        strcpy((km+contatore)->identificativo, request.identificativo);
+        (km+contatore)->key = response.key;
+        (km+contatore)->timestamp = time(NULL);
+        contatore++;
+        if (contatore >= MEMORY_SIZE)
+          contatore = 0;
+        //stampa shared_memory
+        printf("SHARED MEMORY\n");
+        for (int i = 0; i < MEMORY_SIZE; i++) {
+          printf("%i) %s\n", i+1, (km+i)->identificativo);
+          printf("%i) %i\n", i+1, (km+i)->key);
+          printf("%i) %lu\n", i+1, (km+i)->timestamp);
+        }
+      }
+      sendResponse(&request, &response);
+    }
+  } while(bR != -1);
+  free_shared_memory(km);
+  remove_shared_memory(shmid);
 }
